+++ src/main/java/org/jbei/ice/lib/entry/sequence/SequenceDAO.java
@@ -151,9 +151,9 @@ public class SequenceDAO extends HibernateRepository<Sequence> {
         Session session = newSession();
         try {
 
-            Integer itemCount = (Integer) session.createCriteria(Sequence.class)
-                                                 .setProjection(Projections.countDistinct("id"))
-                                                 .add(Restrictions.eq("entry", entry)).uniqueResult();
+            Long itemCount = (Long) session.createCriteria(Sequence.class)
+                                           .setProjection(Projections.countDistinct("id"))
+                                           .add(Restrictions.eq("entry", entry)).uniqueResult();
 
             return itemCount.intValue() > 0;
         } catch (HibernateException e) {
